fixed cart

+++ app/src/main/java/iti/mad/marketly/data/model/cart/CartModel.kt
@@ -1,3 +1,3 @@
 package iti.mad.marketly.data.model.cart
 
-data class CartModel(var id: Long=0,var imageURL:String="",var quantity:Int=0,var price:Double=0.0,var title: String="")
+data class CartModel(var id: Long=0,var imageURL:String="",var quantity:Long=0,var price:Double=0.0,var title: String="")
